requirements.txt
streamlit>=1.27.0
fpdf2>=2.7.6
pandas>=1.3.0
python-dateutil>=2.8.2